+++ plugins/catalog-backend/src/permissions/rules/hasAnnotation.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { Entity } from '@backstage/catalog-model';
 import { RESOURCE_TYPE_CATALOG_ENTITY } from '@backstage/plugin-catalog-common';
+import { z } from 'zod';
 import { createCatalogPermissionRule } from './util';
 
 /**
@@ -31,12 +31,19 @@ export const hasAnnotation = createCatalogPermissionRule({
   description:
     'Allow entities which are annotated with the specified annotation',
   resourceType: RESOURCE_TYPE_CATALOG_ENTITY,
-  apply: (resource: Entity, annotation: string, value?: string) =>
+  paramsSchema: z.object({
+    annotation: z.string().describe('Name of the annotation to match on'),
+    value: z
+      .string()
+      .optional()
+      .describe('Value of the annotation to match on'),
+  }),
+  apply: (resource, { annotation, value }) =>
     !!resource.metadata.annotations?.hasOwnProperty(annotation) &&
     (value === undefined
       ? true
       : resource.metadata.annotations?.[annotation] === value),
-  toQuery: (annotation: string, value?: string) =>
+  toQuery: ({ annotation, value }) =>
     value === undefined
       ? {
           key: `metadata.annotations.${annotation}`,