+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/errors
 
+## 1.1.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/types@1.0.0
+
 ## 1.1.2-next.1
 
 ### Patch Changes