+++ plugins/auth-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-auth-node
 
+## 0.2.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/errors@1.1.2-next.2
+
 ## 0.2.6-next.1
 
 ### Patch Changes