+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-bazaar
 
+## 0.1.25-next.2
+
+### Patch Changes
+
+- f7c2855d76: Added a `Overview Card` for either latest or random projects. Changed `ProjectPreview.tsx` so it take `gridSize` and `useTablePagination` as props.
+- c0352bbc69: Link to the user catalog entity of a member
+- Updated dependencies
+  - @backstage/cli@0.20.0-next.2
+  - @backstage/plugin-catalog@1.6.0-next.2
+  - @backstage/plugin-catalog-react@1.2.0-next.2
+  - @backstage/catalog-client@1.1.1-next.2
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/errors@1.1.2-next.2
+
 ## 0.1.25-next.1
 
 ### Patch Changes