+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @backstage/plugin-catalog-import
 
+## 0.9.0-next.2
+
+### Minor Changes
+
+- b2e6cb6acf: **Breaking**
+  Moved the code search for the existing catalog-info.yaml files to the backend from the frontend. It means it will use the configured GitHub integration's credentials.
+
+  Add the following to your `CatalogBuilder` to have the repo URL ingestion working again.
+
+  ```ts
+  // catalog.ts
+  import { GitHubLocationAnalyzer } from '@backstage/plugin-catalog-backend-module-github';
+  ...
+    builder.addLocationAnalyzers(
+      new GitHubLocationAnalyzer({
+        discovery: env.discovery,
+        config: env.config,
+      }),
+    );
+  ...
+  ```
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-common@1.0.7-next.2
+  - @backstage/plugin-catalog-react@1.2.0-next.2
+  - @backstage/catalog-client@1.1.1-next.2
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/errors@1.1.2-next.2
+  - @backstage/integration@1.3.2-next.2
+  - @backstage/integration-react@1.1.5-next.2
+
 ## 0.8.13-next.1
 
 ### Patch Changes