+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-cost-insights
 
+## 0.11.32-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-cost-insights-common@0.1.1
+
 ## 0.11.32-next.1
 
 ### Patch Changes