+++ plugins/jenkins/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-jenkins
 
+## 0.7.9-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.2.0-next.2
+  - @backstage/plugin-jenkins-common@0.1.9-next.2
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/errors@1.1.2-next.2
+  - @backstage/theme@0.2.16
+
 ## 0.7.9-next.1
 
 ### Patch Changes