+++ plugins/scaffolder-backend-module-yeoman/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-scaffolder-backend-module-yeoman
 
+## 0.2.10-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.7.0-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/types@1.0.0
+
 ## 0.2.10-next.1
 
 ### Patch Changes