+++ plugins/shortcuts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-shortcuts
 
+## 0.3.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/types@1.0.0
+
 ## 0.3.2-next.1
 
 ### Patch Changes