+++ plugins/tech-insights-node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-tech-insights-node
 
+## 0.3.5-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-tasks@0.3.6-next.2
+  - @backstage/backend-common@0.15.2-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/types@1.0.0
+  - @backstage/plugin-tech-insights-common@0.2.7-next.2
+
 ## 0.3.5-next.1
 
 ### Patch Changes