+++ plugins/vault-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-vault-backend
 
+## 0.2.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-tasks@0.3.6-next.2
+  - @backstage/backend-common@0.15.2-next.2
+  - @backstage/backend-test-utils@0.1.29-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/errors@1.1.2-next.2
+
 ## 0.2.3-next.1
 
 ### Patch Changes